backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Database (SQLite for MVP, PostgreSQL optional)
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Client
httpx>=0.26.0

# AI
anthropic>=0.18.1
